refactor hash methods on accounts
