+++ account/account.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/starknet.go/curve"
+	"github.com/NethermindEth/starknet.go/hash"
 	"github.com/NethermindEth/starknet.go/rpc"
 	"github.com/NethermindEth/starknet.go/utils"
 )
@@ -132,7 +133,7 @@ func (account *Account) TransactionHashDeployAccount(tx rpc.DeployAccountTxn, co
 	}
 	calldata := []*felt.Felt{tx.ClassHash, tx.ContractAddressSalt}
 	calldata = append(calldata, tx.ConstructorCalldata...)
-	calldataHash, err := computeHashOnElementsFelt(calldata)
+	calldataHash, err := hash.ComputeHashOnElementsFelt(calldata)
 	if err != nil {
 		return nil, err
 	}
@@ -143,7 +144,7 @@ func (account *Account) TransactionHashDeployAccount(tx rpc.DeployAccountTxn, co
 	}
 
 	// https://docs.starknet.io/documentation/architecture_and_concepts/Network_Architecture/transactions/#deploy_account_hash_calculation
-	return calculateTransactionHashCommon(
+	return hash.CalculateTransactionHashCommon(
 		PREFIX_DEPLOY_ACCOUNT,
 		versionFelt,
 		contractAddress,
@@ -164,7 +165,7 @@ func (account *Account) TransactionHashInvoke(tx rpc.InvokeTxnType) (*felt.Felt,
 			return nil, ErrNotAllParametersSet
 		}
 
-		calldataHash, err := computeHashOnElementsFelt(txn.Calldata)
+		calldataHash, err := hash.ComputeHashOnElementsFelt(txn.Calldata)
 		if err != nil {
 			return nil, err
 		}
@@ -173,7 +174,7 @@ func (account *Account) TransactionHashInvoke(tx rpc.InvokeTxnType) (*felt.Felt,
 		if err != nil {
 			return nil, err
 		}
-		return calculateTransactionHashCommon(
+		return hash.CalculateTransactionHashCommon(
 			PREFIX_TRANSACTION,
 			txnVersionFelt,
 			txn.ContractAddress,
@@ -189,7 +190,7 @@ func (account *Account) TransactionHashInvoke(tx rpc.InvokeTxnType) (*felt.Felt,
 			return nil, ErrNotAllParametersSet
 		}
 
-		calldataHash, err := computeHashOnElementsFelt(txn.Calldata)
+		calldataHash, err := hash.ComputeHashOnElementsFelt(txn.Calldata)
 		if err != nil {
 			return nil, err
 		}
@@ -197,7 +198,7 @@ func (account *Account) TransactionHashInvoke(tx rpc.InvokeTxnType) (*felt.Felt,
 		if err != nil {
 			return nil, err
 		}
-		return calculateTransactionHashCommon(
+		return hash.CalculateTransactionHashCommon(
 			PREFIX_TRANSACTION,
 			txnVersionFelt,
 			txn.SenderAddress,
@@ -222,7 +223,7 @@ func (account *Account) TransactionHashDeclare(tx rpc.DeclareTxnType) (*felt.Fel
 			return nil, ErrNotAllParametersSet
 		}
 
-		calldataHash, err := computeHashOnElementsFelt([]*felt.Felt{txn.ClassHash})
+		calldataHash, err := hash.ComputeHashOnElementsFelt([]*felt.Felt{txn.ClassHash})
 		if err != nil {
 			return nil, err
 		}
@@ -231,7 +232,7 @@ func (account *Account) TransactionHashDeclare(tx rpc.DeclareTxnType) (*felt.Fel
 		if err != nil {
 			return nil, err
 		}
-		return calculateTransactionHashCommon(
+		return hash.CalculateTransactionHashCommon(
 			PREFIX_DECLARE,
 			txnVersionFelt,
 			txn.SenderAddress,
@@ -246,7 +247,7 @@ func (account *Account) TransactionHashDeclare(tx rpc.DeclareTxnType) (*felt.Fel
 			return nil, ErrNotAllParametersSet
 		}
 
-		calldataHash, err := computeHashOnElementsFelt([]*felt.Felt{txn.ClassHash})
+		calldataHash, err := hash.ComputeHashOnElementsFelt([]*felt.Felt{txn.ClassHash})
 		if err != nil {
 			return nil, err
 		}
@@ -255,7 +256,7 @@ func (account *Account) TransactionHashDeclare(tx rpc.DeclareTxnType) (*felt.Fel
 		if err != nil {
 			return nil, err
 		}
-		return calculateTransactionHashCommon(
+		return hash.CalculateTransactionHashCommon(
 			PREFIX_DECLARE,
 			txnVersionFelt,
 			txn.SenderAddress,
@@ -394,4 +395,28 @@ func (account *Account) TransactionByBlockIdAndIndex(ctx context.Context, blockI
 
 func (account *Account) TransactionByHash(ctx context.Context, hash *felt.Felt) (rpc.Transaction, error) {
 	return account.provider.TransactionByHash(ctx, hash)
+}
+
+/*
+Formats the multicall transactions in a format which can be signed and verified by the network and OpenZeppelin account contracts
+*/
+func FmtCalldata(fnCalls []rpc.FunctionCall) []*felt.Felt {
+	callArray := []*felt.Felt{}
+	callData := []*felt.Felt{new(felt.Felt).SetUint64(uint64(len(fnCalls)))}
+
+	for _, tx := range fnCalls {
+		callData = append(callData, tx.ContractAddress, tx.EntryPointSelector)
+
+		if len(tx.Calldata) == 0 {
+			callData = append(callData, &felt.Zero, &felt.Zero)
+			continue
+		}
+
+		callData = append(callData, new(felt.Felt).SetUint64(uint64(len(callArray))), new(felt.Felt).SetUint64(uint64(len(tx.Calldata))+1))
+		callArray = append(callArray, tx.Calldata...)
+	}
+	callData = append(callData, new(felt.Felt).SetUint64(uint64(len(callArray)+1)))
+	callData = append(callData, callArray...)
+	callData = append(callData, new(felt.Felt).SetUint64(0))
+	return callData
 }
